% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata2r.R
\name{s2r}
\alias{s2r}
\title{Translate Stata to R}
\usage{
s2r(input)
}
\arguments{
\item{input}{Enter code as you would in Stata, surrounded by single quotes.}
}
\description{
This function translates common Stata instructions into R and executes the resulting code. Stata instructions are recorded as a string, which are broken down and interpreted.
}
\examples{
s2r('replace x=5 if y==3')
}
\keyword{stata}
