% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata2r.R
\name{stata2r}
\alias{stata2r}
\title{Translate Stata to R}
\usage{
stata2r(code)
}
\arguments{
\item{code}{Enter code as you would in Stata, surrounded by single quotes.}
}
\description{
This function translates common Stata instructions into R and executes the resulting code. Stata instructions are recorded as a string, which are broken down and interpreted.
}
\examples{
stata2r('pwd')

}
\keyword{stata}
